+++ sdk/src/main/java/com/akkaserverless/javasdk/lowlevel/ReplicatedEntityHandler.java
@@ -14,9 +14,12 @@
  * limitations under the License.
  */
 
-package com.akkaserverless.javasdk.replicatedentity;
+package com.akkaserverless.javasdk.lowlevel;
 
 import com.akkaserverless.javasdk.Reply;
+import com.akkaserverless.javasdk.replicatedentity.CommandContext;
+import com.akkaserverless.javasdk.replicatedentity.ReplicatedEntity;
+import com.akkaserverless.javasdk.replicatedentity.StreamedCommandContext;
 import com.google.protobuf.Any;
 
 /**
