+++ sdk/src/test/scala/com/akkaserverless/javasdk/impl/action/AnnotationBasedActionSupportSpec.scala
@@ -29,13 +29,15 @@ import org.scalatest.BeforeAndAfterAll
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.matchers.should.Matchers
-
 import java.util.{Base64, Optional}
 import java.util.concurrent.{CompletableFuture, CompletionStage, TimeUnit}
+
 import scala.compat.java8.FutureConverters._
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
+
 import com.akkaserverless.javasdk.actionspec.ActionspecApi
+import com.akkaserverless.javasdk.lowlevel.ActionHandler
 import com.google.protobuf.ByteString
 
 class AnnotationBasedActionSupportSpec extends AnyWordSpec with Matchers with BeforeAndAfterAll {
